data_ingestion/requirements.txt
psycopg2-binary
python-dotenv
sentence-transformers
numpy # Often a dependency